backend/requirements.txt
flask
python-chess
flask_socketio
eventlet